Fixing the reservations route
