+++ app/forms/reservation_form.py
@@ -1,11 +1,11 @@
 from flask_wtf import FlaskForm
 from wtforms import (
-  DateTimeLocalField, IntegerField, SubmitField )
-from wtforms.validators import DataRequired, Length
+  IntegerField, DateTimeField, SubmitField )
+from wtforms.validators import DataRequired
 
 class ReservationForm(FlaskForm):
     # Users can post one picture that is an image URL right now
     # In the near future, this will be an AWS S3 URL
   num_people = IntegerField('Name', validators=[DataRequired()])
-  date_time = DateTimeLocalField('Which date is your favorite?', format='%m/%d/%y', validators=[DataRequired()])
+  date_time = DateTimeField('Reservation time', format='%Y-%m-%d %H:%M:%S', validators=[DataRequired()])
   submit = SubmitField('Submit')
